refactor: put window frame values in one view model
The shadow margin and the corner radius describe the same rectangle,
but they sat in two different view models, so the window had to reach
through the app content just to learn how far the surface extends past
the frame. With both on the window frame view model the window reads
them straight from the window utility, and the app view model is back
to app state only.

The frame and the surface are easy to mix up, so the glossary now names
the frame and sketches how the two rectangles line up.
