+++ pyproject.toml
@@ -269,8 +269,8 @@ files = [
     "mpvqc/viewmodels/utility/fonts.py",
     "mpvqc/viewmodels/utility/label_width_calculator.py",
     "mpvqc/viewmodels/utility/theme.py",
+    "mpvqc/viewmodels/utility/window_frame.py",
     "mpvqc/viewmodels/utility/window_properties.py",
-    "mpvqc/viewmodels/utility/window_radius.py",
     "mpvqc/viewmodels/views/__init__.py",
     "mpvqc/viewmodels/views/footer.py",
     "mpvqc/viewmodels/views/header/__init__.py",
