+++ test/viewmodels/test_app.py
@@ -10,25 +10,10 @@ from PySide6.QtCore import Qt
 
 from mpvqc.services import (
     KeyCommandGeneratorService,
-    MainWindowService,
-    PlatformService,
     PlayerService,
     SettingsService,
 )
 from mpvqc.viewmodels import MpvqcAppViewModel
-
-
-@pytest.fixture
-def platform_service_mock():
-    return MagicMock(spec_set=PlatformService)
-
-
-@pytest.fixture
-def main_window_service_mock():
-    mock = MagicMock(spec_set=MainWindowService)
-    mock.is_fullscreen = False
-    mock.is_maximized = False
-    return mock
 
 
 @pytest.fixture
@@ -44,15 +29,11 @@ def command_generator_mock():
 @pytest.fixture(autouse=True)
 def configure_inject(
     common_bindings_with,
-    platform_service_mock,
-    main_window_service_mock,
     player_service_mock,
     command_generator_mock,
     settings_service,
 ):
     def custom_bindings(binder: inject.Binder):
-        binder.bind(PlatformService, platform_service_mock)
-        binder.bind(MainWindowService, main_window_service_mock)
         binder.bind(PlayerService, player_service_mock)
         binder.bind(KeyCommandGeneratorService, command_generator_mock)
         binder.bind(SettingsService, settings_service)
@@ -64,15 +45,6 @@ def configure_inject(
 def view_model() -> MpvqcAppViewModel:
     # noinspection PyCallingNonCallable
     return MpvqcAppViewModel()
-
-
-def test_shadow_margin_forwards_service_value(main_window_service_mock):
-    main_window_service_mock.shadow_margin = 128
-
-    # noinspection PyCallingNonCallable
-    view_model = MpvqcAppViewModel()
-
-    assert view_model.shadowMargin == 128
 
 
 def test_forward_key_to_player_sends_generated_command(view_model, player_service_mock, command_generator_mock):
